renamed the modules and tested the notebook
