+++ dnn/sinkhorn_nn.py
